Chat & Builder apps (#68)
# Summary 

This PR begins the revamp of the current admin app experience. We are
deprecating the current combined app and first starting with a
standalone "chat" app, which is intended for users to iterate and add
verified queries to an existing semantic model.

Eventually, we will reincorporate the from-scratch builder flow, but for
now we are starting with the iteration step.

Follow [the branch
README.md](https://github.com/Snowflake-Labs/semantic-model-generator/blob/c997edac7877233f1c25e0a32265b89cc8e2d318/admin_apps/README.md)
to get started and try those apps.

# To-dos

- [x] **Common 'chrome' for both apps.** Maybe there should be some
light and consistent chrome on those two apps just to make them look
coherent. For now, the builder app has logo and title, chat app has
nothing. Maybe there's a middle ground!
- [x] **Make sure the env. variables instructions are clear.** In both
apps, we kinda require that env variables are properly set. It's
stressed in the repo's README, laybe we should add more instructions in
this README too.
- [x] **Make sure the back-end actually works.** I think the two apps
have the UI almost done - but there's a significant part of plugging
wires that must be done. Supposedly, all utils should be found in the
main package, or already in the `shared_utils.py`.
- [x] **Separate Streamlit apps requirements from main package.** For
now, the app requirements are being consolidated in the `pyproject.toml`
which really is about the semantic-model-generator package, and, to me,
should not include the underlying streamlit apps requirements.
- [ ] **Update apps requirements.** I think the current stable streamlit
build is currently missing `streamlit-monaco` (the code editor) and
could almost afford to switch from `streamlit-nightly` to
`streamlit==1.36.0` once ready.
- [ ] **Add tests.** Streamlit has introduced a way to [test Streamlit
apps](https://docs.streamlit.io/develop/api-reference/app-testing), that
could be a neat way to ensure we don't break things in the app!

---------

Co-authored-by: Arnaud ❄️ <arnaudmiribel@gmail.com>
Co-authored-by: Chris Nivera <chris.nivera@snowflake.com>
Co-authored-by: Renee Huang <renee.huang@snowflake.com>