Trigger lint/test workflows on additional files (#87)
The `lint` workflow should run when changing any admin_app code. This
wasn't running because I renamed the directory.

Furthermore, I forgot to regen the `poetry.lock` file when editing
`pyproject.toml`, and this wasn't caught because the build + test action
only runs on the `semantic_model_generator/` dir. Including the toml
file now to catch these omissions in the future. Since we don't have
tests in `admin_apps/` for now we don't need to run the test script for
that directory, but eventually we should.