Assert environment variables upon starting (#98)
Addresses
https://github.com/Snowflake-Labs/semantic-model-generator/issues/92.

It's easy to overlook the required environment variable setup to run
this app. Additionally, it was a bit confusing that in the UI we showed
the option to edit the credentials, but this didn't actually take
effect.

Since these environment variables are unlikely to change significantly
run to run, I'm removing them entirely from the UI, and adding a step on
app startup that asserts that all of the required vars are set. If one
is missing, there is a more clear error shown and instructions provided.

![CleanShot 2024-07-26 at 13 48
23@2x](https://github.com/user-attachments/assets/2c8b6e24-cd98-4ad9-a484-68908b4d1fec)


## Testing

Omit one of the required environment variables from the [setup
instructions](https://github.com/Snowflake-Labs/semantic-model-generator?tab=readme-ov-file#setup),
and observe that you see a popup.