Reuse Snowflake connection throughout Streamlit app (#99)
Resolves
https://github.com/Snowflake-Labs/semantic-model-generator/issues/90,
https://github.com/Snowflake-Labs/semantic-model-generator/issues/97

In the current version of the admin app, we manually instantiate a
`SnowflakeConnection` multiple times throughout the codebase. This isn't
particularly a problem when the user is using standard username/pw
credentials; however, when using MFA or external-browser authentication,
this causes multiple pings/browser windows to pop up, which interrupts
the UX significantly.

I've modified the Streamlit app to instantiate a single top-level
Snowflake connection and cached it via `st.cache_resource`. With
Streamlit's caching, whenever the decorated function is run for the very
first time in the lifetime of the app, the connection is created and
stored in cache. Whenever the function is invoked again, the connection
is retrieved from cache instead of instantiated from scratch. See
https://docs.streamlit.io/develop/concepts/architecture/caching#stcache_resource
for more details.

I also had to make some changes to the underlying model
generation/validation logic so that a connection can optionally be
passed in instead of instantiating new connections.


## Testing

Use the app with a user/account that uses the `externalbrowser`
authenticator type. Verify that you see a single auth popup on app
start, and only once. Verify that you do not see any more popups while
using the app.