Use `union` syntax for type hints (#113)
I had added a few type hints using `|` (e.g. `str | bool`), but didn't
realize this only works for Python >= 3.10. Users that were using 3.9
(which is allowed by our dependency manager) were running into errors
when using the app because the | operand was not recognized.

Changing the app to use `Union` instead of `|` for compatibility.