Dropdown selectors for iteration flow (#121)
Replaces the free text entry fields in the iteration flow with
dropdowns, so that the user can click to select their file instead of
having to type everything in manually. This fully completes #94 .

## Testing
Go through the iteration flow and verify functionality of the dropdown
selectors. Proceed to use the app and upload your YAML to a stage.

Additionally, go through the builder flow, and verify that the step
where you select the Snowflake stage at the end works with the new
dropdown selectors.