Support .env files (#110)
Resolves #107 

Instead of having to export/set the variables in the user's shell, it'd
be handy to support a `.env` file where the prerequisite variables could
be set. We still support setting the variables directly in the shell /
Python code.

This PR adds the `python-dotenv` package to support .env files, and
clarifies the instructions in the README.

## Testing
Set the environment variables in a `.env` file and verify that they are
reflected properly in the app.