Fix rerun bug when saving semantic models (#125)
Some users reported that saving a semantic model would immediately
result in an error that seemed to be Streamlit-specific:
`RerunData(page_script_hash=...)`

It seems that there is a bug with the version of Streamlit that we are
using where `st.rerun()` is throwing a generic Exception - this is in
theory fine behavior since `rerun()` needs to throw an exception to halt
the process, but this causes issues when being called within a
`try/except Exception...` block.

This has been patched by Streamlit but not pushed to a new release yet;
additionally, upgrading would require a few more refactors. For now I'm
simply fixing this by moving the rerun() outside of the try/except
block.