Consolidate admin app dependencies into `pyproject.toml` (#150)
We'd like to revamp the open source app a bit, namely the setup process.
This PR takes the first step into simplifying the setup instructions;
right now, we currently have a `pyproject.toml` corresponding to the CLI
dependencies and a separate`requirements.txt` for the admin Streamlit
app. Since the Streamlit app will be the primarily supported version, we
can bring the admin app dependencies into the main `pyproject.toml` and
delete the former.

Most of the diff is regenerating `pyproject.toml`. Version changes in
the TOML are to bring them in line with `requirements.txt`.