Mark `looker_sdk` as an optional dependency (#151)
Some users may not want to run through partner flows, in which case
installing the Looker SDK is unnecessary. Following the discussion in
https://github.com/Snowflake-Labs/semantic-model-generator/pull/150#issuecomment-2340560372,
this PR moves `looker-sdk` into the "extra" dependencies section,
meaning it won't be installed unless specified (e.g. `poetry install -E
looker`).

To support this, I made some small refactors to only import looker
conditionally. The popup will now be triggered if the user attempts to
click the "Start with partner semantic model" button.

Shown below is a video showing that without looker installed, the rest
of the app is functional, and a warning pops up if a user attempts to
use the Looker integration without the SDK.


https://github.com/user-attachments/assets/5703dc60-948c-4b14-b94d-fa368637790f