Clarify sample value selector during semantic model creation (#161)
The previous behavior of the sample value dropdown was a bit confusing.

Old behavior: depending on the column type, we take
`max(hardcoded_column_sample_value_ct, user input)` to determine the
number of sample values fetched. This was a bit misleading because a
user would select "1" but would still see 25 sample values for certain
columns, since 25 was hardcoded.

New behavior: sample value dropdown is the source of truth; we remove
the hardcoded values so that whatever is selected is the actual maximum
we use for each column. The default number of sample values fetched is
now 3.

See in the screenshots below that selecting "3" now applies the maximum
properly.
 
![CleanShot 2024-09-13 at 13 18
08](https://github.com/user-attachments/assets/f903b480-9d74-4f77-8228-acbe12d440ed)

![CleanShot 2024-09-13 at 13 18
34](https://github.com/user-attachments/assets/ed5ed9ff-b3bb-466e-b2ab-2f48b2d4df19)