Add skeleton joins section on model creation (#163)
We are gating join support behind an "experimental features" checkbox in
the UI. When enabled, we will allow for new features such as joins to be
supported in the semantic model generator. Note that users must have the
prerequisite parameters enabled in their Snowflake accounts for this to
properly work.

For now, we'd like to expose the joins via a commented out section to
simply inform the user of the option. We will quickly follow up by
building an actual UI for selecting tables and their join keys.

![CleanShot 2024-09-17 at 11 09
53@2x](https://github.com/user-attachments/assets/9ab37533-65b9-4810-9b8e-bbd89f9f08a4)

![CleanShot 2024-09-17 at 11 11
07@2x](https://github.com/user-attachments/assets/05c4e688-014f-4192-89ad-87b1354cc402)