Remove analyst generation comment from verified queries (#168)
Queries generated by Analyst include a `--Generated by Cortex Analyst`
comment for attribution. When directly adding these as verified queries
using the OSS, the comment was being left in. When the Analyst then took
into account these verified queries to generate subsequent SQL, the
comment was being preserved and added twice (once because the LLM
preserved the original comment from the verified query, and again by the
Analyst code to append this comment to any SQL queries generated), which
looks strange.

To solve this we are removing the comment before marking the query as a
verified query so that Analyst will not encounter it again.




https://github.com/user-attachments/assets/ab63dbc1-8a0c-4ed9-9b55-5b5c44630435