Populate `primary_key` field for joined tables (#188)
Joins now require a `primary_key` to be specified on the table objects
being used in the joins. When using the join editor, we now attempt to
populate the `primary_key` field on the table by using a `show primary
keys in table...` query. If this query returns no results, we simply set
an empty string column for the user to fill out.