Call Analyst to perform validation (#187)
We currently maintain copies of the validation logic in both the
internal Analyst codepaths as well as this OSS app. Often, the OSS app
can become out of date. Instead of performing validation locally, we
will simply call Analyst with the current YAML string, as it performs
validation at inference time. Any error returned is shown to the user.

The diff for this PR seems big but it's mostly deleting unnecessary code
+ tests.