Allow verified queries to be marked for onboarding (#192)
Users can classify verified queries to be used for onboarding, which
means that they can be returned as predefined questions for a user
onboarding experience. For full detail, see
https://docs.snowflake.com/en/user-guide/snowflake-cortex/cortex-analyst/suggested-questions-feature.

This PR adds the `use_as_onboarding_question` field to the proto as well
as surfaces a simple checkbox UI for admins to mark questions for
onboarding.